Rolled the xml file back to simpler save as df
the current XML parameter functions didn’t simplify the thing at all,
so I rolled back to the old version.

The idea I had originally was that the c++ model gets an xml file, and
gets the parameters from there, so that we don’t have to redo all the
time all the r/c interface. This wasn’t implemented though, how it is
it is just more work