Fixed phylogeny ... but needs CHECKING!!!
This solves the problem that the prune phylogeny function used to
destroy the full phylogeny for creating the pruned phylogeny … before,
this was no problem as we didn’t need it any more, but now we do.

I am not 100% confident that this is fully working. Need to check with
the debugger