Changed main package documentation from .Rd file to .R file
I didn't inlcude
@aliases phylosim-package
@doctype package
as it is stated in the DHARMA package, since this throws the warning "@doctpye  is deprecated".
I used
@keywords internal
"_PACKAGE_"
since that seems to be the new standard and unlike to tahminas post it doesnt seem to add a  automated documention from the description. It seems to still rely on the information from the .R file
Or maybe I'm understanding the idea behind using
@aliases phylosim-package
@doctype package
wrong.
Do we need to have also an documentation when calling "?PhyloSim-package" ?